refactor(webui): extract Theme Resolver into entities/renderer/theme
projectTheme.ts(122줄)를 책임 단위로 분해해 theme/ 모듈로 옮긴다.

- theme/validate.ts: validateTheme + isPlainObject/pickTokens helper.
  invalid shape 거부 + console.warn.
- theme/resolve.ts: resolveThemeVars. light/dark/single-mode/forceScheme
  분기.
- theme/tokens.ts: TOKEN_TO_CSS, TOKEN_KEYS 공유 상수.
- theme/index.ts: 두 함수 외부 export.

entities/renderer/index.ts의 re-export 위치만 ./theme/index.js로 변경.
소비자 import 경로 zero diff. 기존 projectTheme.ts 삭제.

issue #144 Phase F3.
