+++ apps/webui/src/client/entities/renderer/theme/index.ts
@@ -1,0 +1,2 @@
+export { resolveThemeVars } from "./resolve.js";
+export { validateTheme } from "./validate.js";
