+++ apps/webui/src/client/entities/renderer/theme/validate.ts
@@ -1,0 +1,60 @@
+import type { RendererTheme, RendererThemeTokens } from "../renderer.types.js";
+import { TOKEN_KEYS } from "./tokens.js";
+
+function isPlainObject(v: unknown): v is Record<string, unknown> {
+  return typeof v === "object" && v !== null && !Array.isArray(v);
+}
+
+function pickTokens(src: Record<string, unknown>): RendererThemeTokens {
+  const out: RendererThemeTokens = {};
+  for (const key of TOKEN_KEYS) {
+    const value = src[key];
+    if (typeof value === "string" && value.length > 0) {
+      out[key] = value;
+    }
+  }
+  return out;
+}
+
+/**
+ * 렌더러 모듈의 `theme` export를 런타임 검증한다.
+ * 잘못된 shape이면 null 반환 (console.warn 남김).
+ */
+export function validateTheme(raw: unknown): RendererTheme | null {
+  if (raw === undefined || raw === null) return null;
+  if (!isPlainObject(raw)) {
+    console.warn("[renderer.theme] expected an object, got", typeof raw);
+    return null;
+  }
+  if (!isPlainObject(raw.base)) {
+    console.warn("[renderer.theme] missing or invalid `base` tokens");
+    return null;
+  }
+  const base = pickTokens(raw.base);
+  if (Object.keys(base).length === 0) {
+    console.warn("[renderer.theme] `base` contains no recognized tokens");
+    return null;
+  }
+  const theme: RendererTheme = { base };
+
+  if (raw.dark !== undefined) {
+    if (isPlainObject(raw.dark)) {
+      const dark = pickTokens(raw.dark);
+      if (Object.keys(dark).length > 0) {
+        theme.dark = dark;
+      }
+    } else {
+      console.warn("[renderer.theme] `dark` must be an object; ignored");
+    }
+  }
+
+  if (raw.prefersScheme === "light" || raw.prefersScheme === "dark") {
+    theme.prefersScheme = raw.prefersScheme;
+  } else if (raw.prefersScheme !== undefined) {
+    console.warn(
+      '[renderer.theme] `prefersScheme` must be "light" or "dark"; ignored',
+    );
+  }
+
+  return theme;
+}
